backend/requirements.txt
Flask==2.0.1
web3==6.0.0
python-dotenv==0.19.0
flask-sqlalchemy==2.5.1
gunicorn==20.1.0
flask-cors==3.0.10
ruff==0.1.9
black==23.12.1
werkzeug==2.0.3
